Fix inconsistent counter cache behaviour when optimistic locking enabled
If optimistic locking is enabled on a model which `belongs_to`
some associate model, with `counter_cache: true`, the counter cache column
on the associate record is not updated when destroying the belonging
model.
This behaviour was caused by inconsistent behaviour between the private
`#destroy_row` methods in `Locking::Optimistic` and `CounterCache`.

Now both modules always call on `super` to ensure that any other
`#destroy_row` in the chain is also called.
`Persistence#destroy_row` will only call `_delete_row` if there is not
already data on the number of rows affected by delete
(`@_destroy_row_affected_rows`).
`Locking::Optimistic` will set this value because it is responsible for
doing the actual delete when locking is enabled