make acts_like behavior more intuitive
Previously, defining an acts_like-*? method would result in acts_like?
returning true for that type no matter what the underlying method
returned. This change makes the result of acts_like? more intuitive:
the return value of acts_like_*? is respected.

The old functionality is now deprecated, and the new behavior is opt in
based on a configuration value. The deprecation warning will be shown if
the return value of acts_like_* is non-boolean, and acts_like? will
still return true to remain backwards compatible.