Revises examples in the callbacks guide [skip ci]
As a rule of thumb, models are not supposed to be sending emails. Typically, in
a simple application, controller actions are the ones responsible for gluing
things together.

Ideally, creating a user in the console, runner, or test suite, should not
generate an email. That typically happens on a "sign up", which is a higher
level concept. Again, generally speaking.

As any rule of thumb, you may find exceptions, use always your best judgement.
This does not mean sending an email from a callback is categorically wrong, it
just means it is not something you'd generally want to do.

Documentation examples should be a model for readers. Better pick different,
more neutral examples here.

Thanks to @solnic for commenting this example existed.