Document Active Storage attachment and attachments field generators [ci skip]
Found while debugging issue: #42380

I was under an assumption that we have to manually add
`has_one_attached :<name>` and `has_many_attached :<name>`
in the model.

In issue #42296, user was trying to do something similar.

This PR documents the attachment and attachments field
generator to make it more visible to all people.