Disable automatic write protection on replicas
This allows users to use write protection if they want, but no longer treats it as a catch all. Write protection is not accurate enough to classify all cases correctly. Instead users should configure their replicas to not allow writes and rely on the database erroring if the query is not allowed.

Co-authored-by: Alex Ghiculescu <alexghiculescu@gmail.com>