Add Authenticated Controllers example to Active Storage guide [skip ci]
The default Active Storage controllers are publicly accessible by
default.  The generated URLs are hard to guess, but anyone that knows
the blob URL will be able to download it.

The guides warn about this and advice implementing your own
authenticated controllers if required. But currently there is no
guidance on how to implement this.

This commit adds an example, based on the following comment:
https://github.com/rails/rails/pull/38843#issuecomment-619083761

The original warning mentioned 'the URLs being public'. To avoid
confusion with the S3 URL, which are only public for 5 minutes, the
warning has be changed to 'the controllers being public'.

The warning has also been moved up, to make it clear it applies to all
redirect and proxy URLs.