Allow preloads on instance dependent associations
Instance dependent associations are associations which are defined with
a scope taking an argument which makes the scope they are selected from
dependent on the owner record.

This presented a challenge for preloading and eager_loading, since each
record's association being loaded could require a different scope.
Previously this raised an exception if attempted.

This commit adds the ability to preload instance dependent scopes (eager
loading is still unsupported). This is done by detecting instance
dependent scopes in the preloader and building the scope and
Preloader::Association for each record. Because we now merge similar
scopes later in the preloading process, any records which end up with
the same instance dependent scope will be loaded from a single query.

In the case that there's a different scope for each record, this will
perform N+1 queries. I think it's up to the app developer to consider
this (this is essentially the same as preloading polymorphic scopes).

Co-authored-by: John Crepezzi <john.crepezzi@gmail.com>
Co-authored-by: Adam Hess <HParker@github.com>
Co-authored-by: Eileen M. Uchitelle <eileencodes@gmail.com>
Co-authored-by: Dinah Shi <dinahshi@github.com>