Forced encoding for deterministic encryption and other improvements (#42491)
* Fix: re-encrypting will preserve case when ignore_case is used

The implementation was outdated in relation to the new approach where we
override accessors.

* Add support for exists?(...) when querying encrypted attributes

* Add a new option to force encoding for deterministic attributes

Different encodings result in different ciphertexts. For
deterministically encrypted attributes, this can result in having
attributes with the same value that fails to satisfy uniqueness
constraints due to having different encodings.

This adds a new option `forced_encoding_for_deterministic_encryption:`
that will be UTF-8 by default. User can disabled this new behavior by
setting the option to nil.

* Add changelog entry