Add clear message when rails new finishes
Running `rails new` for the first time on a computer can easily generate
hundreds of lines of output. Before this change, the end of the output
would read:

    Webpacker successfully installed 🎉 🍰

While that line makes sense when all of the output is read from top to
bottom, it is kinda confusing for a new user who just created a new
Rails app and glossed over all of the intermediate output from the
command.

This change adds an explicit message to the end of `rails new` that
says 1) the command finished successfully and 2) here's some more info
if you don't know what to do next.