Expand guides descriptions for the index
Most of the current descriptions for the guides on the index page are
very sparse. They don't describe the subjects that are handled in the guides.
You have to know the frameworks to know what to look for.
If we expand the descriptions, it can help beginners find the correct
guide.

This commit copies some of the introductions from the introductions of
the guides themselves to the index descriptions. Some descriptions were
expanded to be more informative.

Co-authored-by: Sunny Ripert <sunny@sunfox.org>