Make `FlashHash#delete` behave like `Hash#delete`
Prior to this commit, the `FlashHash#delete` return value was the
instance itself.

[The last time this code was changed][76c2ea7] (10 years ago at the time
of submission), `FlashHash` defined its own delegating versions of
`Hash` methods, as an alternative to inheriting from `Hash` directly.

This behavior isn't mentioned in the method's documentation (this commit
also introduces a documentation comment for the method), and isn't
covered by the test suite.

This behavior diverges from [Hash#delete][], since that method returns
the delete value(s).

[Hash#delete]: https://ruby-doc.org/core-3.0.1/Hash.html#method-i-delete
[FlashHash#delete]: https://edgeapi.rubyonrails.org/classes/ActionDispatch/Flash/FlashHash.html#method-i-delete
[76c2ea7]: https://github.com/rails/rails/commit/76c2ea7882a83159408bdf1f7c363f442a65c4f1