Cancel current postgres query on rollback
If a postgres query inside a transaction is interrupted, the query keeps
executing in the database but control flow is returned to ruby. When we
attempt to rollback the transaction, it blocks until the previous query
finishes.

Since we're going to rollback anyway, we just cancel the previous query
before issuing the rollback. Cancelling when no query is executing is
just a no-op.