Improve setting custom loggers to LogSubscribers
Any subclass of `ActiveSupport::LogSubscriber` may have a custom logger
set to it, without altering the one set on the base class or any other
subclasses. If one is not set, it still defaults to
`ActiveSupport::LogSubscriber.logger`, which in turn defaults to
`Rails.logger` when in Rails.

This was achieved by:

- calling `self.class.logger` instead of `LogSubscriber.logger` in
  `#logger`
- defaulting `.logger` on any subclass to the one from its superclass
- removed memoization from `.logger` to prevent memoizing a default
  from its superclass before it's been assigned