Add '80 character wrap' to API documentation guidelines
Currently there is no documentation we can point to that prescribes
wrapping documentation at 80 characters. Adding it to the API guidelines
helps contributors improve their PR's and allows reviewers to point to
the guides.

The 80 character limit is defined in the "Reviewing issues and PRs"
Basecamp document and is occaisionally mentioned in reviews, for example
the following comment by Eileen:

> All new guides should be wrapped to 80 characters.

https://github.com/rails/rails/pull/37885#discussion_r400931085