"Middleware#remove" is renamed "Middleware#delete!"
This commit intends to clarify the difference between
`Middleware#delete` and `Middleware#delete!`.

The former method silently fails when the target item is
not found, while the latter raises an error.

The functionality of `delete!` has been introduced in 688ed70
and given a name `remove`. This commit only renames it.

Also, a brief description of `delete!` method is now
provided for guides so that users can acknowledge the difference.