Update `ActiveRecord::Batches#in_batches` to make use of `ActiveRecord::Base.implicit_order_column` when present.
When using `ActiveRecord::Batches#in_batches` ActiveRecord sorts records by
primary key. This causes unpredictable behavior and errors when the primary
key is not an auto-incrementing data type, like a UUID. This change follows up
on similar changes made to the +first+ and +last+ finder calls and changes
`ActiveRecord::Batches#in_batches` to have it sort by the `implicit_order_column`
when that attribute has been configured. This allows for the batch processing of
records that do not have an auto-incrementing primary key.

The only slight difference in behavior comes when using batches with an
`implicit_order_column` that is non-unique. The batch size will represent the
number of unique values to be processed in the batch not the number of records.
For Example:

  If the `implicit_order_column` has values of:
  ```
  [:a, :b, :c, :c, :d, :e, :f]
  ```
  And the user calls `Model.in_batches(of: 3)` the first batch will contain four
  records, `[:a, :b, :c, :c]`, and the second batch will contain the three remaining
  records, `[:d, :e, :f]`. No records will be skipped, barring any race conditions, but
  the number of records processed may vary between batches.