feat: Allow specifying a tag that isnt just `form`
# Purpose:

When using <https://shoelace.style> I wanted to hook in and create my own form builder. The problem is that there is no nice way to override the base form_with to use a different element (in this case `<sl-form>`)

## What this does

This allows a user calling `form_with` to specify an HTML tag to use a different form tag.

## Example

```rb
<%= form_with model: Model, html: { tag: "sl-form" } do |f| %>
<% end %>
```

will then generate the following html:

```html
<sl-form tag="sl-form"></sl-form>
```

instead of the typical `<form></form>`

## Questions / Considerations

Im not sure if theres a more semantic or appropriate way to make this happen as the `tag` attribute will appear on the parent form, but I figured I could submit this small PR and at least start a discussion. I apologize if this is outside of Rails / the form builder.