Make SafeBuffer#split return Array of SafeBuffers
When calling `#split` on an instance of `ActiveSupport::SafeBuffer`,
return an Array of instances of `ActiveSupport::SafeBuffer`.

If the original string instance has been marked as safe, splitting it
into its parts should not be a mutative action. Therefore the
safe-marking should also apply to the new instances, similarly to how
using [SafeBuffer#slice][] preserves the safety.

[SafeBuffer#slice]: https://github.com/rails/rails/blob/c3518b4a9c9cbd5fc264231c7b153ac6f6195019/activesupport/test/safe_buffer_test.rb#L200-L204