improve PG index schema queries
this is safe because it's impossible to create an index in a schema besides the the schema that the table is in. it's far more performant if you have many schemas/tables in a single database because pg_class has an index on (relname, relnamespace), whereas pg_index has no such index.