Set record_timestamps = true in AR::InternalMetadata
Fix: https://github.com/rails/rails/issues/42983

The usefulness of these timestamps is a bit questionable,
but I suppose they could be useful for debugging, and
removing them would require to ship a migration etc.

Since `record_timestamps` is now a `class_attribute` we
can simply force it to `true` in just that class.