Allow Postgres JSON operator to be used in order
Previously, any JSON operator would return the error
`ActiveRecord::UnknownAttributeReference`. This error is intended to
protect against SQL injection. It requires the developer to wrap the
query in `Arel.sql` to avoid the error. Referencing a JSON column should
be treated the same as referencing a string column.

This commit only allows `->>` to be used in order.