Enable passing env var paths to cli -m option
Currently, if you add the --template/-m option to a railsrc file, you
lose the ability to define the path to your rails template using an
environment variable, which you have when passing that option from the
command line.

This patch adds that ability by shelling out to interpret the passed
path.